DjangoOfficeProject/requirements.txt
# Django 核心框架
django==5.2.6

# 数据库相关
mysqlclient==2.2.0  # MySQL数据库驱动

# REST API框架
djangorestframework==3.15.2  # Django REST Framework

# 认证相关
djangorestframework-simplejwt==5.3.1  # JWT认证

# 跨域支持
corsheaders==4.3.0  # CORS中间件

# 异步任务处理
celery==5.4.0  # Celery任务队列
redis==5.0.1  # Redis客户端

gevent==23.9.1  # Gevent协程库，用于提高Celery并发性能

# 缓存相关
django-redis==5.4.0  # Django Redis缓存后端

# 工具库
python-dotenv==1.0.0  # 环境变量管理